Release v1.1.0

+++ driven/storage/adapter.go
@@ -1,41 +1,44 @@
-/*
- *   Copyright (c) 2020 Board of Trustees of the University of Illinois.
- *   All rights reserved.
-
- *   Licensed under the Apache License, Version 2.0 (the "License");
- *   you may not use this file except in compliance with the License.
- *   You may obtain a copy of the License at
-
- *   http://www.apache.org/licenses/LICENSE-2.0
-
- *   Unless required by applicable law or agreed to in writing, software
- *   distributed under the License is distributed on an "AS IS" BASIS,
- *   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- *   See the License for the specific language governing permissions and
- *   limitations under the License.
- */
+// Copyright 2022 Board of Trustees of the University of Illinois.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
 
 package storage
 
 import (
 	"fmt"
-	"github.com/rokwire/core-auth-library-go/tokenauth"
-	"go.mongodb.org/mongo-driver/bson"
-	"go.mongodb.org/mongo-driver/bson/bsontype"
-	"go.mongodb.org/mongo-driver/bson/primitive"
-	"go.mongodb.org/mongo-driver/mongo/options"
 	"log"
 	"polls/core/model"
+	"polls/driven/groups"
 	"strconv"
 	"time"
+
+	"go.mongodb.org/mongo-driver/bson"
+	"go.mongodb.org/mongo-driver/bson/primitive"
+	"go.mongodb.org/mongo-driver/mongo/options"
 )
 
 const (
-	statusCreated    = "created"
-	statusStarted    = "started"
-	statusTerminated = "terminated"
-	settingsKey      = "stadium"
-	eventInterval    = 100 * time.Millisecond
+	// PollStatusCreated status created
+	PollStatusCreated = "created"
+
+	// PollStatusStarted status started
+	PollStatusStarted = "started"
+
+	// PollStatusTerminated status terminated
+	PollStatusTerminated = "terminated"
+
+	settingsKey   = "stadium"
+	eventInterval = 100 * time.Millisecond
 )
 
 // Adapter implements the Storage interface
@@ -69,10 +72,10 @@ func NewStorageAdapter(config *model.Config) *Adapter {
 }
 
 // GetPolls retrieves all polls with an ability to filter
-func (sa *Adapter) GetPolls(user *tokenauth.Claims, filter model.PollsFilter, filterByToMembers bool) ([]model.Poll, error) {
+func (sa *Adapter) GetPolls(user *model.User, filter model.PollsFilter, filterByToMembers bool, membership *groups.GroupMembership) ([]model.Poll, error) {
 	mongoFilter := bson.D{}
 	if user != nil {
-		mongoFilter = append(mongoFilter, primitive.E{Key: "org_id", Value: user.OrgID})
+		mongoFilter = append(mongoFilter, primitive.E{Key: "org_id", Value: user.Claims.OrgID})
 	}
 
 	if len(filter.PollIDs) > 0 {
@@ -87,23 +90,17 @@ func (sa *Adapter) GetPolls(user *tokenauth.Claims, filter model.PollsFilter, fi
 
 	if filter.MyPolls != nil && *filter.MyPolls == true && filter.RespondedPolls != nil && *filter.RespondedPolls == true {
 		mongoFilter = append(mongoFilter, primitive.E{Key: "$or", Value: []primitive.M{
-			{"poll.userid": user.Subject},
-			{"responses.userid": user.Subject},
+			{"poll.userid": user.Claims.Subject},
+			{"responses.userid": user.Claims.Subject},
 		}})
 	} else {
 		if filter.MyPolls != nil && *filter.MyPolls == true {
-			mongoFilter = append(mongoFilter, primitive.E{Key: "poll.userid", Value: user.Subject})
+			mongoFilter = append(mongoFilter, primitive.E{Key: "poll.userid", Value: user.Claims.Subject})
 		}
 
 		if filter.RespondedPolls != nil && *filter.RespondedPolls == true {
-			mongoFilter = append(mongoFilter, primitive.E{Key: "responses.userid", Value: user.Subject})
+			mongoFilter = append(mongoFilter, primitive.E{Key: "responses.userid", Value: user.Claims.Subject})
 		}
-	}
-
-	if filter.GroupPolls != nil && *filter.GroupPolls {
-		mongoFilter = append(mongoFilter, primitive.E{Key: "poll.group_id", Value: bson.M{"$exists": true}})
-		mongoFilter = append(mongoFilter, primitive.E{Key: "poll.group_id", Value: bson.M{"$ne": bsontype.Null}})
-		mongoFilter = append(mongoFilter, primitive.E{Key: "poll.group_id", Value: bson.M{"$ne": ""}})
 	}
 
 	if filter.Pin != nil {
@@ -119,11 +116,21 @@ func (sa *Adapter) GetPolls(user *tokenauth.Claims, filter model.PollsFilter, fi
 	}
 
 	if filterByToMembers {
+		var innerFilter primitive.M
+		if membership != nil && len(membership.GroupIDsAsAdmin) > 0 {
+			innerFilter = primitive.M{"$or": []primitive.M{
+				primitive.M{"poll.group_id": bson.M{"$in": membership.GroupIDsAsAdmin}},
+				primitive.M{"poll.to_members.user_id": user.Claims.Subject},
+			}}
+		} else {
+			innerFilter = primitive.M{"poll.to_members.user_id": user.Claims.Subject}
+		}
+
 		mongoFilter = append(mongoFilter, primitive.E{Key: "$or", Value: []primitive.M{
 			primitive.M{"poll.to_members": primitive.Null{}},
 			primitive.M{"poll.to_members": primitive.M{"$exists": true, "$size": 0}},
-			primitive.M{"poll.to_members.user_id": user.Subject},
-			primitive.M{"poll.user_id": user.Subject},
+			primitive.M{"poll.user_id": user.Claims.Subject},
+			innerFilter,
 		}})
 	}
 
@@ -147,12 +154,31 @@ func (sa *Adapter) GetPolls(user *tokenauth.Claims, filter model.PollsFilter, fi
 }
 
 // GetPoll retrieves a single poll
-func (sa *Adapter) GetPoll(user *tokenauth.Claims, id string) (*model.Poll, error) {
+func (sa *Adapter) GetPoll(user *model.User, id string, filterByToMembers bool, membership *groups.GroupMembership) (*model.Poll, error) {
 
 	if objID, err := primitive.ObjectIDFromHex(id); err == nil {
 		filter := bson.D{
-			primitive.E{Key: "org_id", Value: user.OrgID},
+			primitive.E{Key: "org_id", Value: user.Claims.OrgID},
 			primitive.E{Key: "_id", Value: objID},
+		}
+
+		if filterByToMembers {
+			var innerFilter primitive.M
+			if membership != nil && len(membership.GroupIDsAsAdmin) > 0 {
+				innerFilter = primitive.M{"$or": []primitive.M{
+					primitive.M{"poll.group_id": bson.M{"$in": membership.GroupIDsAsAdmin}},
+					primitive.M{"poll.to_members.user_id": user.Claims.Subject},
+				}}
+			} else {
+				innerFilter = primitive.M{"poll.to_members.user_id": user.Claims.Subject}
+			}
+
+			filter = append(filter, primitive.E{Key: "$or", Value: []primitive.M{
+				primitive.M{"poll.to_members": primitive.Null{}},
+				primitive.M{"poll.to_members": primitive.M{"$exists": true, "$size": 0}},
+				primitive.M{"poll.user_id": user.Claims.Subject},
+				innerFilter,
+			}})
 		}
 
 		var list []model.Poll
@@ -174,11 +200,11 @@ func (sa *Adapter) GetPoll(user *tokenauth.Claims, id string) (*model.Poll, erro
 }
 
 // CreatePoll creates a poll
-func (sa *Adapter) CreatePoll(user *tokenauth.Claims, poll model.Poll) (*model.Poll, error) {
-	poll.OrgID = user.OrgID
+func (sa *Adapter) CreatePoll(user *model.User, poll model.Poll) (*model.Poll, error) {
+	poll.OrgID = user.Claims.OrgID
 	poll.ID = primitive.NewObjectID()
-	poll.UserID = user.Subject
-	poll.UserName = user.Name
+	poll.UserID = user.Claims.Subject
+	poll.UserName = user.Claims.Name
 	poll.DateCreated = time.Now()
 
 	_, err := sa.db.polls.InsertOne(poll)
@@ -191,12 +217,13 @@ func (sa *Adapter) CreatePoll(user *tokenauth.Claims, poll model.Poll) (*model.P
 }
 
 // UpdatePoll updates a poll
-func (sa *Adapter) UpdatePoll(user *tokenauth.Claims, poll model.Poll) (*model.Poll, error) {
+func (sa *Adapter) UpdatePoll(user *model.User, poll model.Poll) (*model.Poll, error) {
+
 	if len(poll.ID) > 0 {
 
 		poll.DateUpdated = time.Now().UTC()
 		filter := bson.D{
-			primitive.E{Key: "org_id", Value: user.OrgID},
+			primitive.E{Key: "org_id", Value: user.Claims.OrgID},
 			primitive.E{Key: "_id", Value: poll.ID},
 		}
 
@@ -228,17 +255,17 @@ func (sa *Adapter) UpdatePoll(user *tokenauth.Claims, poll model.Poll) (*model.P
 }
 
 // StartPoll starts an existing poll
-func (sa *Adapter) StartPoll(user *tokenauth.Claims, pollID string) error {
+func (sa *Adapter) StartPoll(user *model.User, pollID string) error {
 
-	poll, err := sa.GetPoll(user, pollID)
+	poll, err := sa.GetPoll(user, pollID, true, nil)
 	if err != nil {
 		fmt.Printf("error storage.Adapter.StartPoll(%s) - %s", pollID, err)
 		return fmt.Errorf("error storage.Adapter.StartPoll(%s) - %s", pollID, err)
 	}
 
 	if poll != nil {
-		if poll.Status != statusStarted {
-			poll.Status = statusStarted
+		if poll.Status != PollStatusStarted {
+			poll.Status = PollStatusStarted
 			_, err = sa.UpdatePoll(user, *poll)
 			if err != nil {
 				fmt.Printf("error storage.Adapter.StartPoll(%s) - %s", pollID, err)
@@ -253,16 +280,16 @@ func (sa *Adapter) StartPoll(user *tokenauth.Claims, pollID string) error {
 }
 
 // EndPoll ends an existing poll
-func (sa *Adapter) EndPoll(user *tokenauth.Claims, pollID string) error {
-	poll, err := sa.GetPoll(user, pollID)
+func (sa *Adapter) EndPoll(user *model.User, pollID string) error {
+	poll, err := sa.GetPoll(user, pollID, true, nil)
 	if err != nil {
 		fmt.Printf("error storage.Adapter.EndPoll(%s) - %s", pollID, err)
 		return fmt.Errorf("error storage.Adapter.EndPoll(%s) - %s", pollID, err)
 	}
 
 	if poll != nil {
-		if poll.Status != statusTerminated {
-			poll.Status = statusTerminated
+		if poll.Status != PollStatusTerminated {
+			poll.Status = PollStatusTerminated
 			_, err = sa.UpdatePoll(user, *poll)
 			if err != nil {
 				fmt.Printf("error storage.Adapter.EndPoll(%s) - %s", pollID, err)
@@ -277,23 +304,25 @@ func (sa *Adapter) EndPoll(user *tokenauth.Claims, pollID string) error {
 }
 
 // DeletePoll deletes a poll
-func (sa *Adapter) DeletePoll(user *tokenauth.Claims, id string) error {
-	filter := bson.D{
-		primitive.E{Key: "org_id", Value: user.OrgID},
-		primitive.E{Key: "_id", Value: id},
-	}
-	_, err := sa.db.polls.DeleteOne(filter, nil)
-	if err != nil {
-		fmt.Printf("error storage.Adapter.DeletePoll(): error while delete poll (%s) - %s", id, err)
-		return fmt.Errorf("error storage.Adapter.DeletePoll(): error while delete poll (%s) - %s", id, err)
-	}
+func (sa *Adapter) DeletePoll(user *model.User, id string) error {
+	if objID, err := primitive.ObjectIDFromHex(id); err == nil {
+		filter := bson.D{
+			primitive.E{Key: "org_id", Value: user.Claims.OrgID},
+			primitive.E{Key: "_id", Value: objID},
+		}
+		_, err := sa.db.polls.DeleteOne(filter, nil)
+		if err != nil {
+			fmt.Printf("error storage.Adapter.DeletePoll(): error while delete poll (%s) - %s", id, err)
+			return fmt.Errorf("error storage.Adapter.DeletePoll(): error while delete poll (%s) - %s", id, err)
+		}
 
+	}
 	return nil
 
 }
 
 // VotePoll votes a poll
-func (sa *Adapter) VotePoll(user *tokenauth.Claims, pollID string, vote model.PollVote) error {
+func (sa *Adapter) VotePoll(user *model.User, pollID string, vote model.PollVote) error {
 
 	if objID, err := primitive.ObjectIDFromHex(pollID); err == nil {
 		now := time.Now().UTC()
@@ -341,9 +370,161 @@ func (m *database) onDataChanged(changeDoc map[string]interface{}) {
 	coll := nsMap["coll"]
 
 	record := changeDoc["fullDocument"]
-	recordMap := record.(map[string]interface{})
+	var recordMap map[string]interface{}
+	if record != nil {
+		recordMap = record.(map[string]interface{})
+	}
 
 	if m.listener != nil {
 		m.listener.OnCollectionUpdated(coll.(string), recordMap)
 	}
+}
+
+// GetSurvey retrieves a single survey
+func (sa *Adapter) GetSurvey(user *model.User, id string) (*model.Survey, error) {
+
+	filter := bson.M{"_id": id, "org_id": user.Claims.OrgID, "app_id": user.Claims.AppID}
+	var entry model.Survey
+	err := sa.db.surveys.FindOne(filter, &entry, nil)
+	if err != nil {
+		fmt.Printf("error storage.Adapter.GetSurvey(%s) - %s", id, err)
+		return nil, fmt.Errorf("error storage.Adapter.GetSurvey(%s) - %s", id, err)
+	}
+
+	return &entry, nil
+}
+
+// CreateSurvey creates a poll
+func (sa *Adapter) CreateSurvey(survey model.Survey) (*model.Survey, error) {
+
+	_, err := sa.db.surveys.InsertOne(survey)
+	if err != nil {
+		fmt.Printf("error storage.Adapter.CreateSurvey(%s) - %s", survey.ID, err)
+		return nil, fmt.Errorf("error storage.Adapter.CreateSurvey(%s) - %s", survey.ID, err)
+	}
+
+	return &survey, nil
+}
+
+// UpdateSurvey updates a survey
+func (sa *Adapter) UpdateSurvey(user *model.User, survey model.Survey) error {
+
+	if len(survey.ID) > 0 {
+
+		now := time.Now().UTC()
+		filter := bson.M{"_id": survey.ID, "creator_id": user.Claims.Subject, "org_id": user.Claims.OrgID, "app_id": user.Claims.AppID}
+		update := bson.M{"$set": bson.M{
+			"title":                 survey.Title,
+			"data":                  survey.Data,
+			"scored":                survey.Scored,
+			"result_rule":           survey.ResultRule,
+			"type":                  survey.Type,
+			"stats":                 survey.SurveyStats,
+			"sensitive":             survey.Sensitive,
+			"default_data_key":      survey.DefaultDataKey,
+			"default_data_key_rule": survey.DefaultDataKeyRule,
+			"date_updated":          now,
+		}}
+
+		res, err := sa.db.surveys.UpdateOne(filter, update, nil)
+		if err != nil {
+			fmt.Printf("error storage.Adapter.UpdateSurvey(%s) - %s", survey.ID, err)
+			return fmt.Errorf("error storage.Adapter.UpdateSurvey(%s) - %s", survey.ID, err)
+		}
+		if res.ModifiedCount != 1 {
+			fmt.Printf("storage.Adapter.UpdateSurvey(%s) invalid id", survey.ID)
+			return fmt.Errorf("storage.Adapter.UpdateSurvey(%s) invalid id", survey.ID)
+		}
+
+	}
+
+	return nil
+}
+
+// DeleteSurvey deletes a survey
+func (sa *Adapter) DeleteSurvey(user *model.User, id string) error {
+
+	filter := bson.M{"_id": id, "creator_id": user.Claims.Subject, "org_id": user.Claims.OrgID, "app_id": user.Claims.AppID}
+
+	res, err := sa.db.surveys.DeleteOne(filter, nil)
+	if err != nil {
+		return fmt.Errorf("error storage.Adapter.DeleteSurvey(): error while delete survey (%s) - %s", id, err)
+	}
+	if res.DeletedCount != 1 {
+		fmt.Printf("storage.Adapter.DeleteSurvey(%s) invalid id", id)
+		return fmt.Errorf("storage.Adapter.DeleteSurvey(%s) invalid id", id)
+	}
+
+	return nil
+}
+
+// GetSurveyResponse gets a survey response by ID
+func (sa *Adapter) GetSurveyResponse(user *model.User, id string) (*model.SurveyResponse, error) {
+
+	filter := bson.M{"_id": id, "user_id": user.Claims.Subject, "org_id": user.Claims.OrgID, "app_id": user.Claims.AppID}
+	var entry model.SurveyResponse
+	err := sa.db.surveyResponses.FindOne(filter, &entry, nil)
+	if err != nil {
+		fmt.Printf("error storage.Adapter.GetSurveyResponse(%s) - %s", id, err)
+		return nil, fmt.Errorf("error storage.Adapter.GetSurveyResponse(%s) - %s", id, err)
+	}
+
+	return &entry, nil
+}
+
+// CreateSurveyResponse creates a new survey response
+func (sa *Adapter) CreateSurveyResponse(surveyResponse model.SurveyResponse) (*model.SurveyResponse, error) {
+
+	_, err := sa.db.surveyResponses.InsertOne(surveyResponse)
+	if err != nil {
+		fmt.Printf("error storage.Adapter.CreateSurveyResponse(%s) - %s", surveyResponse.ID, err)
+		return nil, fmt.Errorf("error storage.Adapter.CreateSurveyResponse(%s) - %s", surveyResponse.ID, err)
+	}
+
+	return &surveyResponse, nil
+}
+
+// UpdateSurveyResponse updates an existing service response
+func (sa *Adapter) UpdateSurveyResponse(user *model.User, id string, survey model.Survey) error {
+
+	if len(id) > 0 {
+
+		now := time.Now().UTC()
+		filter := bson.M{"_id": id, "user_id": user.Claims.Subject, "org_id": user.Claims.OrgID, "app_id": user.Claims.AppID}
+		update := bson.M{"$set": bson.M{
+			"survey":       survey,
+			"date_updated": now,
+		}}
+
+		res, err := sa.db.surveyResponses.UpdateOne(filter, update, nil)
+		if err != nil {
+			fmt.Printf("error storage.Adapter.UpdateSurveyResponse(%s) - %s", id, err)
+			return fmt.Errorf("error storage.Adapter.UpdateSurveyResponse(%s) - %s", id, err)
+		}
+		if res.ModifiedCount != 1 {
+			fmt.Printf("storage.Adapter.UpdateSurveyResponse(%s) invalid id", id)
+			return fmt.Errorf("storage.Adapter.UpdateSurveyResponse(%s) invalid id", id)
+		}
+
+	}
+
+	return nil
+}
+
+// DeleteSurveyResponse deletes a survey response
+func (sa *Adapter) DeleteSurveyResponse(user *model.User, id string) error {
+
+	filter := bson.M{"_id": id, "user_id": user.Claims.Subject, "org_id": user.Claims.OrgID, "app_id": user.Claims.AppID}
+
+	res, err := sa.db.surveyResponses.DeleteOne(filter, nil)
+	if err != nil {
+		fmt.Printf("error storage.Adapter.DeleteSurveyResponse(%s) - %s", id, err)
+		return fmt.Errorf("error storage.Adapter.DeleteSurveyResponse(): error while delete survey response (%s) - %s", id, err)
+	}
+	if res.DeletedCount != 1 {
+		fmt.Printf("storage.Adapter.DeleteSurveyResponse(%s) invalid id", id)
+		return fmt.Errorf("storage.Adapter.DeleteSurveyResponse(%s) invalid id", id)
+	}
+
+	return nil
 }
